refactor (DiscountPolicy) : 할인 정책을 discount 패키지 내부에서 분리
- christmas.domain.discount의 해당 클래스가 많아져 구분하기 위해 분리함
