+++ src/main/java/christmas/domain/discount/policy/WeekdayDiscount.java
@@ -1,18 +1,18 @@
-package christmas.domain.discount;
+package christmas.domain.discount.policy;
 
 import christmas.domain.plan.Plan;
 
-public class WeekendDiscount extends DiscountPolicy {
+public class WeekdayDiscount extends DiscountPolicy {
 
     private static final int DISCOUNT_AMOUNT_PER_MENU = 2_023;
 
     @Override
     boolean isSatisfyPrecondition(Plan plan) {
-        return plan.isWeekend();
+        return plan.isWeekday();
     }
 
     @Override
     int calculateDiscountAmount(Plan plan) {
-        return DISCOUNT_AMOUNT_PER_MENU * plan.countMainMenu();
+        return DISCOUNT_AMOUNT_PER_MENU * plan.countDessertMenu();
     }
 }
