+++ src/test/java/christmas/domain/badge/BadgeTest.java
@@ -2,7 +2,6 @@ package christmas.domain.badge;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import christmas.domain.badge.Badge;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
