+++ src/test/java/christmas/domain/discount/policy/MockDiscountPolicy.java
@@ -1,4 +1,4 @@
-package christmas.domain.discount;
+package christmas.domain.discount.policy;
 
 import christmas.domain.plan.Plan;
